orc/requirements.txt
google-adk
a2a-sdk
httpx
python-dotenv
fastapi
uvicorn
pydantic
sqlalchemy
